% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_funs.R
\name{crop_images}
\alias{crop_images}
\title{Crop images
Detects field of view using Hough Circle Transform and crops images at the
edge.}
\usage{
crop_images(path, path_out, radius = c(2200L, 2400L), crop_factor = 0.9)
}
\arguments{
\item{path}{Path with directory containing image files.}

\item{path_out}{Path to write cropped image files.}

\item{radius}{Vector with minimum and maximum radius of edge in pixels.
Defaults to \code{c(2200,2400)}.}

\item{crop_factor}{Multiplier for crop radius. Defaults to \code{0.9}.}
}
\description{
Crop images
Detects field of view using Hough Circle Transform and crops images at the
edge.
}
