% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_zs_dataframe.R
\name{expand_zs_dataframe}
\alias{expand_zs_dataframe}
\title{Expand data.frame}
\usage{
expand_zs_dataframe(
  df,
  c_path,
  c_id,
  c_start,
  c_end,
  extension = "JPG",
  digits = 4
)
}
\arguments{
\item{df}{A data.frame containing the first and last photo, file path, and
grouping variable.}

\item{c_path}{String or numerical index specifying the column where the file
paths can be found. Paths should}

\item{c_id}{String or numerical index specifying column where factor can
be found for grouping images.}

\item{c_start}{String or numerical index specifying the column containing
the first photo of each stack.}

\item{c_end}{String or numerical index specifying the column containing
the last photo of each stack.}

\item{extension}{String specifying the file extension of the images. The
extension must be in the correct case so it matches exactly to the extension
of the files to be stacked.}

\item{digits}{How many digits should be in the number.}
}
\value{
Expanded data.frame
}
\description{
This function facilitates stacking photos that are numbered sequentially. It
must be supplied with a \code{data.frame} with columns containing the
information about the images to be stacked.
}
