% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_funs.R
\name{image_similarity}
\alias{image_similarity}
\title{Image similarity
Calculates structural similarity score for consecutive images in the provided
directory.}
\usage{
image_similarity(path_in, path_out, ext = "jpg")
}
\arguments{
\item{path_in}{Path with directory containing image files.}

\item{path_out}{Path to write the similarities}

\item{ext}{File extension. Defaults to "jpg".}
}
\description{
Image similarity
Calculates structural similarity score for consecutive images in the provided
directory.
}
