% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_zs_batch.R
\name{stack_files}
\alias{stack_files}
\title{Stack files in folders}
\usage{
stack_files(
  df,
  c_path,
  c_split,
  temp = TRUE,
  show_progress = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{df}{data.frame containing at least 2 columns containing the paths of
the files to parse and a grouping factor.}

\item{c_path}{String or numerical index specifying column where paths can be found.}

\item{c_split}{String or numerical index specifying column where factor can
be found for grouping images.}

\item{temp}{Logical. If TRUE, the function will stack files into a temp
directory which will be deleted when the operation is completed. Thus,
the organization of the original image files will be maintained. Otherwise,
images will be moved permanently into new folders according to the provided
grouping variable. Defaults to TRUE.}
}
\description{
Stack files in folders
}
\author{
Ethan Bass
}
