% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_stacks.R
\name{identify_stacks}
\alias{identify_stacks}
\title{Identify stack breakpoints}
\usage{
identify_stacks(
  path,
  cutoff = 0.9,
  check_lt = 13,
  cutoff2 = 0.85,
  interactive = FALSE,
  check_gt = 25,
  ext = "jpg",
  verbose = getOption("verbose")
)
}
\arguments{
\item{path}{Directory containing images.}

\item{cutoff}{Similarity threshold.}

\item{check_lt}{Minimum number of images in stack to trigger secondary check.}

\item{cutoff2}{Secondary similarity threshold for secondary check.}

\item{interactive}{Boolean. Whether to do interactive checking. Requires
specification of \code{check_gt}.}

\item{check_gt}{Maximum number of images to trigger secondary check when
\code{verbose} is \code{TRUE}.}

\item{ext}{File extension. Defaults to \code{"jpg"}.}

\item{verbose}{Boolean. Whether to print verbose output to console.}
}
\description{
Identify stack breakpoints
}
