% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_zs_batch.R
\name{run_zs_batch}
\alias{run_zs_batch}
\title{create and run batchscript in zerene stacker}
\usage{
run_zs_batch(
  files,
  c_path = 1,
  c_split = 2,
  path_out,
  stacker = c("pmax", "dmap"),
  temp,
  path_template = NULL,
  path_xml = NULL,
  stack = TRUE
)
}
\arguments{
\item{files}{A data.frame or character vector. If \code{stack == TRUE},
a \code{data.frame} should be provided with at least 2 columns containing the paths of
the files to parse (\code{c_path}) and a grouping factor (\code{c_split}). If
\code{stack == FALSE}, a character vector should be provided containing paths
to the stacked files.}

\item{c_path}{String or numerical index specifying column where paths can be found.}

\item{c_split}{String or numerical index specifying column where factor can
be found for grouping images}

\item{path_out}{directory to export converted files.}

\item{stacker}{Which stacking algorithm to use. Either \code{pmax} or \code{dmap}.}

\item{temp}{Logical. If TRUE, the function will stack files into a temp
directory which will be deleted when the operation is completed. Thus,
the organization of the original image files will be maintained. Otherwise,
images will be moved into folders by stack.}

\item{path_template}{Path to custom template to be customized according to
the provided arguments (optional).}

\item{path_xml}{path to write xml file (optional).}

\item{stack}{Logical. Whether to stack the files or not. Defaults to TRUE.}
}
\value{
If \code{return_paths} is TRUE, the function will return a vector of paths to the newly created files.
If \code{return_paths} is FALSE and \code{export_format} is \code{csv}, the function will return a list
of chromatograms in \code{data.frame} format. Otherwise, it will not return anything.
}
\description{
create and run batchscript in zerene stacker
}
\section{Side effects}{
 Chromatograms will be exported in the format specified
by \code{export_format} in the folder specified by \code{path_out}.
}

\author{
Ethan Bass
}
